requirements.txt
asgiref==3.7.2
astor==0.8.1
attrs==23.1.0
bandit==1.7.5
black==23.7.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.2
darglint==1.8.1
defusedxml==0.7.1
Django==4.2.3
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
docutils==0.20.1
drf-yasg==1.21.6
eradicate==2.3.0
filetype==1.2.0
flake8==6.0.0
flake8-bandit==4.1.1
flake8-broken-line==1.0.0
flake8-bugbear==23.7.10
flake8-commas==2.1.0
flake8-comprehensions==3.14.0
flake8-debugger==4.1.2
flake8-docstrings==1.7.0
flake8-eradicate==1.5.0
flake8-isort==6.0.0
flake8-quotes==3.3.2
flake8-rst-docstrings==0.3.0
flake8-string-format==0.3.0
gitdb==4.0.10
GitPython==3.1.32
idna==3.4
inflection==0.5.1
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mccabe==0.7.0
mdurl==0.1.2
mypy-extensions==1.0.0
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==23.1
pathspec==0.11.1
pbr==5.11.1
pep8-naming==0.13.3
platformdirs==3.9.1
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pycparser==2.21
pydocstyle==6.3.0
pyflakes==3.0.1
Pygments==2.15.1
PyJWT==2.8.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
pyVoIP==1.6.5
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
restructuredtext-lint==1.4.0
rich==13.4.2
simplejson==3.19.1
smmap==5.0.0
snowballstemmer==2.2.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
stevedore==5.1.0
tomli==2.0.1
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.3
wemake-python-styleguide==0.18.0
